+++ core/src/main/kotlin/org/radarbase/core/model/project/participant/ProjectParticipant.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2025 King's College London
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.radarbase.core.model.project.participant
+
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonElement
+
+/**
+ * Shared serializable ProjectParticipant model.
+ *
+ * Originated from project-service; enriched with optional [metadata]
+ * field used by delegate-api.
+ */
+@Serializable
+data class ProjectParticipant(
+    val participantId: String,
+    val projectId: String,
+    val status: String,
+    val enrollmentDate: String? = null,
+    val attributes: Map<String, String> = emptyMap(),
+    val metadata: Map<String, JsonElement> = emptyMap(),
+)
+
