+++ project-service/src/main/kotlin/org/radarbase/project/service/ProjectService.kt
@@ -3,7 +3,7 @@ package org.radarbase.project.service
 import org.radarbase.core.model.project.Group
 import org.radarbase.core.model.project.Project
 import org.radarbase.project.model.CreateGroupRequest
-import org.radarbase.project.model.ProjectParticipant
+import org.radarbase.core.model.project.participant.ProjectParticipant
 import org.radarbase.project.model.RadarProject
 
 interface ProjectService {
